Improve genlib database randomness
dds uses mt random number generator that has very long period but
doesn't use array based back to increase random state between boards.
That combination limits number starting sequences to 2^32 that is limit
for 32 bit input seed. But long period makes each generate completely
different boards.

That makes it desirable to request large number of boards from a single
dds execution. This change limits single process to about 52k boards.

Another important factor is to try to avoid same seed to be used for the
generation. To that end it is best that genlib uses random number
generator to advance the seed instead of simple arithmetic increment.
That makes it very unlikely to hit same seed but even if that happens it
is very likely to be only single seed conflict.

Signed-off-by: Pauli Nieminen <suokkos@gmail.com>